# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: PrepVI R1 API
  version: 1.0.0
  description: Same-origin API contract for PrepVI US-01–30. Sensitive fields are intentionally absent from error schemas.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /questions:
    get:
      security: []
      operationId: listQuestions
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: topic, in: query, schema: { type: string } }
        - { name: difficulty, in: query, schema: { $ref: "#/components/schemas/Difficulty" } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses: { "200": { $ref: "#/components/responses/QuestionPage" } }
  /questions/{questionId}:
    parameters: [ { $ref: "#/components/parameters/QuestionId" } ]
    get:
      security: []
      operationId: getQuestion
      responses: { "200": { $ref: "#/components/responses/Question" }, "404": { $ref: "#/components/responses/Error" } }
  /practice-progress/{questionId}:
    parameters: [ { $ref: "#/components/parameters/QuestionId" } ]
    put:
      operationId: updatePracticeProgress
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /admin/questions:
    get:
      operationId: adminListQuestions
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses: { "200": { $ref: "#/components/responses/QuestionPage" } }
    post:
      operationId: createQuestion
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "201": { $ref: "#/components/responses/Question" }, "422": { $ref: "#/components/responses/Error" } }
  /admin/questions/{questionId}:
    parameters: [ { $ref: "#/components/parameters/QuestionId" } ]
    put:
      operationId: updateQuestion
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Question" }, "409": { $ref: "#/components/responses/Error" } }
  /admin/questions/{questionId}/lifecycle:
    parameters: [ { $ref: "#/components/parameters/QuestionId" } ]
    patch:
      operationId: changeQuestionLifecycle
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Question" }, "409": { $ref: "#/components/responses/Error" } }
...
  /admin/taxonomy/versions:
    post:
      operationId: createTaxonomyVersion
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "201": { $ref: "#/components/responses/Success" } }
  /admin/taxonomy/versions/{versionId}:
    parameters: [ { name: versionId, in: path, required: true, schema: { type: string, format: uuid } } ]
    patch:
      operationId: updateTaxonomyVersion
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
  /admin/topics:
    post:
      operationId: createTopic
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "201": { $ref: "#/components/responses/Success" } }
  /admin/topics/{itemId}:
    parameters: [ { name: itemId, in: path, required: true, schema: { type: string, format: uuid } } ]
    patch:
      operationId: updateTopic
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
  /admin/positions:
    post:
      operationId: createPosition
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "201": { $ref: "#/components/responses/Success" } }
  /admin/positions/{itemId}:
    parameters: [ { name: itemId, in: path, required: true, schema: { type: string, format: uuid } } ]
    patch:
      operationId: updatePosition
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "200": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
  /admin/topic-aliases:
    post:
      operationId: createTopicAlias
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "201": { $ref: "#/components/responses/Success" } }
  /admin/topic-aliases/{aliasId}:
    parameters: [ { name: aliasId, in: path, required: true, schema: { type: string, format: uuid } } ]
    delete:
      operationId: deleteTopicAlias
      requestBody: { $ref: "#/components/requestBodies/JsonInput" }
      responses: { "204": { description: Alias deleted } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: __Host-prepvi_session }
  parameters:
...
    QuestionId: { name: questionId, in: path, required: true, schema: { type: string } }
...
  requestBodies:
    JsonInput:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  schemas:
    Recovery:
      type: object
      required: [kind, retryable, retryAfterSeconds]
      properties:
        kind: { type: string, enum: [RETRY_SAFE, EDIT_INPUT, REUPLOAD, PASTE_TEXT, SELECT_ANOTHER_SLOT, WAIT, CONTACT_SUPPORT, NONE] }
        retryable: { type: boolean }
        retryAfterSeconds: { type: [integer, "null"] }
    Error:
      type: object
      additionalProperties: false
      required: [code, message, correlationId, fieldErrors, recovery]
      properties:
        code: { type: string }
        message: { type: string }
        correlationId: { type: string, format: uuid }
        fieldErrors: { type: object, additionalProperties: { type: string } }
        recovery: { $ref: "#/components/schemas/Recovery" }
...
    Difficulty: { type: string, enum: [EASY, MEDIUM, HARD] }
...
    Question:
      type: object
      required: [id, slug, title, content, difficulty, lifecycleStatus, topics, positions, version]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        title: { type: string }
        content: { type: string }
        answerCriteria: { type: array, items: { type: string } }
        difficulty: { $ref: "#/components/schemas/Difficulty" }
        lifecycleStatus: { type: string, enum: [DRAFT, IN_REVIEW, PUBLISHED, ARCHIVED] }
        topics: { type: array, items: { type: string } }
        positions: { type: array, items: { type: string } }
        version: { type: integer }
...
    PageInfo:
      type: object
      required: [page, pageSize, total]
      properties: { page: { type: integer }, pageSize: { type: integer }, total: { type: integer } }
  responses:
    Success: { description: Successful resource response, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
...
    Error: { description: Safe actionable error, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
    Question: { description: Question resource, content: { application/json: { schema: { $ref: "#/components/schemas/Question" } } } }
    QuestionPage: { description: Question page, content: { application/json: { schema: { type: object, required: [items, pageInfo], properties: { items: { type: array, items: { $ref: "#/components/schemas/Question" } }, pageInfo: { $ref: "#/components/schemas/PageInfo" } } } } } }
...